requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
async-timeout==4.0.2
atlastk==0.13
attrs==21.4.0
ccxt==1.88.95
ccxtpro==1.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
expiringdict==1.2.1
frozenlist==1.3.0
idna==3.3
iso8601==1.0.2
krakenex==2.1.0
multidict==6.0.2
numpy==1.22.3
pandas==1.4.2
pycares==4.2.0
pycparser==2.21
pykrakenapi==0.3.0
pypika-tortoise==0.1.5
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
tortoise-orm==0.19.1
typing-extensions==4.2.0
urllib3==1.26.9
websocket-client==1.3.3
yapf==0.32.0
yarl==1.7.2
